cloud: Goland-move to ./pkg/storage/cloud
Release note: None
